P2/src/DB/requirements.txt
aiofiles==24.1.0
apturl==0.5.2
astroid==2.9.3
async-timeout==5.0.1
asyncpg==0.30.0
attrs==21.2.0
bcrypt==3.2.0
blinker==1.8.2
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
click==8.1.7
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.21
exceptiongroup==1.2.2
fasteners==0.14.1
Flask==3.0.3
Flask-JWT-Extended==4.6.0
future==0.18.2
greenlet==3.1.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httplib2==0.20.2
Hypercorn==0.17.3
hyperframe==6.0.1
idna==3.3
importlib-metadata==4.6.4
isort==5.6.4
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.4
jsonschema==3.2.0
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy-object-proxy==0.0.0
lockfile==0.12.2
logilab-common==1.8.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.5
mccabe==0.6.1
monotonic==1.6
more-itertools==8.10.0
mypy-extensions==0.4.3
netifaces==0.11.0
numpy==1.26.0
oauthlib==3.2.0
olefile==0.46
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
platformdirs==2.5.1
priority==2.0.0
protobuf==3.12.4
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pycairo==1.20.1
pycups==2.0.1
pygame==2.6.0
PyGObject==3.42.1
PyJWT==2.3.0
pylint==2.12.2
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
Quart==0.19.6
reportlab==3.6.8
requests==2.25.1
SecretStorage==3.3.1
six==1.16.0
SQLAlchemy==2.0.36
systemd-python==234
taskgroup==0.0.0a4
texttable==1.6.4
toml==0.10.2
tomli==2.0.1
typing_extensions==4.12.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
wadllib==1.3.6
websocket-client==1.2.3
Werkzeug==3.0.4
wrapt==1.13.3
wsproto==1.2.0
xdg==5
xkit==0.0.0
zipp==1.0.0
